requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
PyYAML>=5.4.0
pytest>=6.2.0
opencv-python>=4.5.0
